+++ tests/test_convert.py
@@ -57,7 +57,8 @@ class Test(unittest.TestCase):
         c = self.pdf.to_csv(precision=3)
         assert c.split("\r\n")[9] == (
             "char,1,45.83,58.826,656.82,674.82,117.18,117.18,135.18,12.996,"
-            '18.0,12.996,,,,,,TimesNewRomanPSMT,,,,"(0, 0, 0)",,,18.0,,,,,Y,,1,'
+            '18.0,12.996,,,,,,TimesNewRomanPSMT,,,"(1, 0, 0, 1, 45.83, 660.69)"'
+            ',,"(0, 0, 0)",,,18.0,,,,,Y,,1,'
         )
 
         io = StringIO()
@@ -111,7 +112,8 @@ class Test(unittest.TestCase):
 
         assert res.decode("utf-8").split("\r\n")[9] == (
             "char,1,45.83,58.826,656.82,674.82,117.18,117.18,135.18,12.996,"
-            '18.0,12.996,,,,,,TimesNewRomanPSMT,,,,"(0, 0, 0)",,,18.0,,,,,Y,,1,'
+            '18.0,12.996,,,,,,TimesNewRomanPSMT,,,"(1, 0, 0, 1, 45.83, 660.69)"'
+            ',,"(0, 0, 0)",,,18.0,,,,,Y,,1,'
         )
 
     def test_page_to_dict(self):